Dedupe ST UnkEntityFunc0
