+++ src/st/no3/41C80.c
@@ -542,30 +542,7 @@ u8 func_801C5668(s32 arg0, s32 arg1) {
 
 #include "../adjust_value_within_threshold.h"
 
-// Duplicate
-void UnkEntityFunc0(u16 slope, s16 speed) {
-    Entity* entity;
-    s32 moveX;
-    s32 moveY;
-
-    moveX = rcos(slope) * speed;
-    entity = g_CurrentEntity;
-
-    if (moveX < 0) {
-        moveX += 15;
-    }
-
-    entity->velocityX = moveX >> 4;
-
-    moveY = rsin(slope) * speed;
-    entity = g_CurrentEntity;
-
-    if (moveY < 0) {
-        moveY += 15;
-    }
-
-    entity->velocityY = moveY >> 4;
-}
+#include "../unk_entity_func0.h"
 
 u16 func_801C5794(s16 arg0, s16 arg1) { return ratan2(arg1, arg0); }
