+++ src/st/wrp/861C.c
@@ -556,25 +556,7 @@ u8 func_8018C004(s32 x, s32 y) {
 
 #include "../adjust_value_within_threshold.h"
 
-void UnkEntityFunc0(u16 slope, s16 speed) {
-    Entity* entity;
-    s32 moveX;
-    s32 moveY;
-
-    moveX = rcos(slope) * speed;
-    entity = g_CurrentEntity;
-    if (moveX < 0) {
-        moveX += 15;
-    }
-    entity->velocityX = moveX >> 4;
-
-    moveY = rsin(slope) * speed;
-    entity = g_CurrentEntity;
-    if (moveY < 0) {
-        moveY += 15;
-    }
-    entity->velocityY = moveY >> 4;
-}
+#include "../unk_entity_func0.h"
 
 u16 func_8018C130(s16 x, s16 y) { return ratan2(y, x); }
